refactor tests + commands
